+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+## v1.2.2 - 2026.02.27
+
+### Changes
+Changed main library from `github.com/jonathonwebb/getopt` to `github.com/spf13/cobra`
+
+Changed a bit of the project's structure to adapt to Golang.
+
 ## v1.2.1 - 2026.02.20
 
+### Changes
 Started Commita's migration from **C** to **Golang**
 
 Created `CHANGELOGS.md`